feat: 회원 의류 삭제 파라미터 타입 변경

+++ src/main/java/com/responseor/ootw/controller/MemberController.java
@@ -71,7 +71,7 @@ public class MemberController {
     }
 
     @GetMapping("/my-clothes/{id}/delete")
-    public ResponseEntity<?> deleteMemberClothes(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteMemberClothes(@PathVariable("id") int id) {
         memberService.deleteMemberClothes(id);
 
         return ResponseEntity.ok(null);
